able to translate string, refactored returns via pointer to heap not onto stack

+++ structs/structs_definitions.h
@@ -23,7 +23,7 @@ typedef struct hashable_string_struct {
 // for returning both pointer to string on heap & its length.
 typedef struct malloc_str {
     char* str_ptr;      // the malloc-ed string.
-    int32_t str_len;    // length of the malloc-ed string.
+    size_t str_len;    // length of the malloc-ed string.
 } malloc_str;
 
 // for parsing user arguments from argv.
